fixing tests

+++ writeopia/src/commonTest/kotlin/io/writeopia/normalization/builder/StepsMapNormalizationBuilderTest.kt
@@ -18,7 +18,7 @@ class StepsMapNormalizationBuilderTest {
 
         assertEquals(
             StoryTypes.GROUP_IMAGE.type,
-            normalized[0]!!.type,
+            normalized[0.0]!!.type,
             "The first non space StoryUnit should be a GroupStep"
         )
 
@@ -29,21 +29,21 @@ class StepsMapNormalizationBuilderTest {
         )
         assertEquals(
             StoryTypes.GROUP_IMAGE.type,
-            normalized[2]?.type,
+            normalized[2.0]?.type,
             "The images in the same position should be merged into GroupImage"
         )
         assertEquals(
             3,
-            normalized[2]?.steps?.size,
+            normalized[2.0]?.steps?.size,
             "The images in the same position should be merged into GroupImage"
         )
         assertEquals(
             StoryTypes.TEXT.type,
-            normalized[normalized.size - 1]?.type,
+            normalized[(normalized.size - 1).toDouble()]?.type,
             "The last message should stay as it was"
         )
 
-        val group = (normalized[4])
+        val group = (normalized[4.0])
         group!!.steps.forEach { storyUnit ->
             assertEquals(
                 group.id,
