+++ writeopia_ui/src/commonTest/kotlin/io/writeopia/ui/backstack/SnapshotBackstackManagerTest.kt
@@ -16,7 +16,7 @@ class SnapshotBackstackManagerTest {
 
     @Test
     fun `when adding the first action - the manager should notify it is possible to revert`() {
-        backstackManager.addTextState(state, 2)
+        backstackManager.addTextState(state, 2.0)
 
         assertEquals(true, backstackManager.canUndo.value)
         assertEquals(false, backstackManager.canRedo.value)
@@ -25,7 +25,7 @@ class SnapshotBackstackManagerTest {
     @Test
     fun `when adding and popping the stack - the manager should notify correctly`() {
         repeat(3) { i ->
-            val state = StoryState(stories = mapOf(i to StoryStep(type = StoryTypes.TEXT.type)))
+            val state = StoryState(stories = mapOf(i.toDouble() to StoryStep(type = StoryTypes.TEXT.type)))
 
             backstackManager.addState(state)
         }
@@ -52,7 +52,7 @@ class SnapshotBackstackManagerTest {
     @Test
     fun `when adding a new story - it should no longer be possible to mover forward`() {
         repeat(3) { i ->
-            val state = StoryState(stories = mapOf(i to StoryStep(type = StoryTypes.TEXT.type)))
+            val state = StoryState(stories = mapOf(i.toDouble() to StoryStep(type = StoryTypes.TEXT.type)))
 
             backstackManager.addState(state)
         }
@@ -67,7 +67,7 @@ class SnapshotBackstackManagerTest {
 
         backstackManager.addState(
             StoryState(
-                stories = mapOf(0 to StoryStep(type = StoryTypes.TEXT.type))
+                stories = mapOf(0.0 to StoryStep(type = StoryTypes.TEXT.type))
             )
         )
 
@@ -77,15 +77,15 @@ class SnapshotBackstackManagerTest {
 
     @Test
     fun `it should be possible to revert a check`() {
-        val state = StoryState(stories = mapOf(0 to StoryStep(type = StoryTypes.TEXT.type)))
+        val state = StoryState(stories = mapOf(0.0 to StoryStep(type = StoryTypes.TEXT.type)))
         backstackManager.addState(state)
 
         val checkState =
-            StoryState(stories = mapOf(0 to StoryStep(type = StoryTypes.CHECK_ITEM.type)))
+            StoryState(stories = mapOf(0.0 to StoryStep(type = StoryTypes.CHECK_ITEM.type)))
         backstackManager.addState(checkState)
 
         val listItemState =
-            StoryState(stories = mapOf(0 to StoryStep(type = StoryTypes.UNORDERED_LIST_ITEM.type)))
+            StoryState(stories = mapOf(0.0 to StoryStep(type = StoryTypes.UNORDERED_LIST_ITEM.type)))
         backstackManager.addState(listItemState)
 
         assertEquals(listItemState, backstackManager.previousState(state))
@@ -95,15 +95,15 @@ class SnapshotBackstackManagerTest {
 
     @Test
     fun `it should be possible to revert and redo a check`() {
-        val state = StoryState(stories = mapOf(0 to StoryStep(type = StoryTypes.TEXT.type)))
+        val state = StoryState(stories = mapOf(0.0 to StoryStep(type = StoryTypes.TEXT.type)))
         backstackManager.addState(state)
 
         val checkState =
-            StoryState(stories = mapOf(0 to StoryStep(type = StoryTypes.CHECK_ITEM.type)))
+            StoryState(stories = mapOf(0.0 to StoryStep(type = StoryTypes.CHECK_ITEM.type)))
         backstackManager.addState(checkState)
 
         val listItemState =
-            StoryState(stories = mapOf(0 to StoryStep(type = StoryTypes.UNORDERED_LIST_ITEM.type)))
+            StoryState(stories = mapOf(0.0 to StoryStep(type = StoryTypes.UNORDERED_LIST_ITEM.type)))
         backstackManager.addState(listItemState)
 
         assertEquals(listItemState, backstackManager.previousState(StoryState(stories = emptyMap())))
